requirements.txt
kafka-python==2.3.0
pandas==2.3.3
requests==2.32.5
psycopg2-binary==2.9.11